Always compile documents three times.
I tried just examining the output of the previous compile run for
"Rerun" (which, by the way, is the most you should check for — because
the surrounding text can differ), but LaTeX won't always tell you when
anothe run is needed.

More specifically, LaTeX will tell you after one run that a rerun is
needed in order for the TOC to get typeset. But on the second run it
doesn't tell you a third run is needed to properly offset the page
numbers in the TOC to take the TOC pages into account.

So, three compile runs it is! Oh well.

Resolves #11.